[refactor] Rename search services to providers and move gateway to dedicated folder
