+++ src/services/algoliaSearchProvider.ts
