+++ src/services/registrySearchProvider.ts
